refactor(grow): move validate_grow_output to grow_validation.py
validate_grow_output was in polish_validation.py for historical
reasons. Move it alongside the rest of GROW's validation code so the
stage owns its own contract. Adds GrowContractError for the
forthcoming stage-exit wiring.

Private helpers (_check_arc_traversal_completeness,
_check_predecessor_cycles, _check_intersection_group_paths) move
with it. POLISH entry contract continues to work via updated import.

Pure refactor — no behavior change.
